# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TranscodeProductResponse": {
        "description": "One product this video has already been transcoded into.\n\n`size_bytes` 是**这次请求当场 stat 出来的**，不是表里的抄本；文件没了就是 `None`。\n`deleted_at` 是\"哪一次核对发现它没了\"，所以它是历史，不是现状 —— 现状只看\n`size_bytes`。",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "output_path": {
            "title": "Output Path",
            "type": "string"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "target_format": {
            "title": "Target Format",
            "type": "string"
          }
        },
        "required": [
          "id",
          "target_format",
          "output_path",
          "size_bytes",
          "created_at",
          "deleted_at"
        ],
        "title": "TranscodeProductResponse",
        "type": "object"
      },
      "TranscodeRequest": {
        "description": "Request model for transcoding.",
        "properties": {
          "target_format": {
            "title": "Target Format",
            "type": "string"
          }
        },
        "required": [
          "target_format"
        ],
        "title": "TranscodeRequest",
        "type": "object"
      },
      "TranscodeResponse": {
        "description": "Response model for transcoding.",
        "properties": {
          "output_path": {
            "title": "Output Path",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "target_format": {
            "title": "Target Format",
            "type": "string"
          },
          "video_id": {
            "title": "Video Id",
            "type": "integer"
          }
        },
        "required": [
          "video_id",
          "status",
          "target_format",
          "output_path"
        ],
        "title": "TranscodeResponse",
        "type": "object"
      },
      "TranscodeStatusResponse": {
        "description": "Response model for transcoding status.",
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "is_transcoding": {
            "title": "Is Transcoding",
            "type": "boolean"
          },
          "output_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Path"
          },
          "progress": {
            "default": 0.0,
            "title": "Progress",
            "type": "number"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "target_format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Format"
          },
          "video_id": {
            "title": "Video Id",
            "type": "integer"
          }
        },
        "required": [
          "video_id",
          "is_transcoding",
          "status"
        ],
        "title": "TranscodeStatusResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Backend API for video management platform",
    "title": "Video Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/transcode/{video_id}": {
      "post": {
        "description": "Start transcoding a video.",
        "operationId": "transcode_video_endpoint_api_transcode__video_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscodeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transcode Video Endpoint",
        "tags": [
          "transcode"
        ]
      }
    },
    "/api/transcode/{video_id}/cancel": {
      "post": {
        "description": "Cancel an active transcoding task.",
        "operationId": "cancel_transcode_api_transcode__video_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Transcode",
        "tags": [
          "transcode"
        ]
      }
    },
    "/api/transcode/{video_id}/outputs": {
      "get": {
        "description": "Products this video has already been transcoded into, newest first.\n\n服务重启之后 `_jobs` 那本账就空了，这一条是产物唯一还在的出处（#154）。",
        "operationId": "list_transcode_products_api_transcode__video_id__outputs_get",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TranscodeProductResponse"
                  },
                  "title": "Response List Transcode Products Api Transcode  Video Id  Outputs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Transcode Products",
        "tags": [
          "transcode"
        ]
      }
    },
    "/api/transcode/{video_id}/status": {
      "get": {
        "description": "Get transcoding status.",
        "operationId": "get_transcode_status_api_transcode__video_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscodeStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Transcode Status",
        "tags": [
          "transcode"
        ]
      }
    },
...
  }
}